requirements.txt
Django==5.2.2
torch==2.5.1+cu121
transformers==4.52.4
langdetect==1.0.9
openai-whisper==20240930
moviepy==2.2.1
accelerate==1.7.0